requirements.txt
asgiref==3.6.0
contourpy==1.0.7
cycler==0.11.0
Django==4.1.7
djangorestframework==3.14.0
fonttools==4.39.0
joblib==1.2.0
kiwisolver==1.4.4
numpy==1.24.2
packaging==23.0
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sqlparse==0.4.3
threadpoolctl==3.1.0
tzdata==2022.7
